@@ -2,35 +2,63 @@
 
 package io.github.hyuwah.draggableviewlib
 
+import android.animation.Animator
+import android.animation.AnimatorListenerAdapter
 import android.graphics.PixelFormat
 import android.os.Build
+import android.util.Log
 import android.view.MotionEvent
 import android.view.View
+import android.view.ViewGroup
 import android.view.WindowManager
 import io.github.hyuwah.draggableviewlib.Draggable.DRAG_TOLERANCE
 import io.github.hyuwah.draggableviewlib.Draggable.DURATION_MILLIS
+//import io.github.hyuwah.draggableviewlib.DraggableView.StickyRestSide
 import kotlin.math.abs
 import kotlin.math.max
 import kotlin.math.min
 import kotlin.math.roundToInt
 
+internal fun View.marginStart(): Float {
+    return ((layoutParams as? ViewGroup.MarginLayoutParams)?.marginStart ?: 0).toFloat()
+}
+
+internal fun View.marginEnd(): Float {
+    return ((layoutParams as? ViewGroup.MarginLayoutParams)?.marginEnd ?: 0).toFloat()
+}
+
+internal fun View.marginTop(): Float {
+    return ((layoutParams as? ViewGroup.MarginLayoutParams)?.topMargin ?: 0).toFloat()
+}
+
+internal fun View.marginBottom(): Float {
+    return ((layoutParams as? ViewGroup.MarginLayoutParams)?.bottomMargin ?: 0).toFloat()
+}
+
 @JvmOverloads
-fun View.makeDraggable(
-    stickyAxis: Draggable.STICKY = Draggable.STICKY.NONE,
+internal fun View.setupDraggable(
+//    minimizeBtnListener: DraggableView.Listener,
+    stickyAxis: DraggableView.Mode = DraggableView.Mode.NON_STICKY,
     animated: Boolean = true,
-    draggableListener: DraggableListener? = null
+    draggableListener: DraggableListener? = null,
 ) {
     var widgetInitialX = 0f
     var widgetDX = 0f
     var widgetInitialY = 0f
     var widgetDY = 0f
+
+    val marginStart = marginStart()
+    val marginTop = marginTop()
+    val marginEnd = marginEnd()
+    val marginBottom = marginBottom()
+
     setOnTouchListener { v, event ->
         val viewParent = v.parent as View
         val parentHeight = viewParent.height
         val parentWidth = viewParent.width
-        val xMax = parentWidth - v.width
+        val xMax = parentWidth - v.width - marginEnd
         val xMiddle = parentWidth / 2
-        val yMax = parentHeight - v.height
+        val yMax = parentHeight - v.height - marginBottom
         val yMiddle = parentHeight / 2
 
         when (event.actionMasked) {
@@ -42,95 +70,121 @@ fun View.makeDraggable(
             }
             MotionEvent.ACTION_MOVE -> {
                 var newX = event.rawX + widgetDX
-                newX = max(0F, newX)
-                newX = min(xMax.toFloat(), newX)
+                newX = max(marginStart, newX)
+                newX = min(xMax, newX)
                 v.x = newX
 
                 var newY = event.rawY + widgetDY
-                newY = max(0F, newY)
-                newY = min(yMax.toFloat(), newY)
+                newY = max(marginTop, newY)
+                newY = min(yMax, newY)
                 v.y = newY
 
                 draggableListener?.onPositionChanged(v)
+//                minimizeBtnListener.onPositionChanged(v, StickyRestSide.HIDE)
             }
             MotionEvent.ACTION_UP -> {
                 when (stickyAxis) {
-                    Draggable.STICKY.AXIS_X -> {
+                    DraggableView.Mode.STICKY_X -> {
                         if (event.rawX >= xMiddle) {
                             if (animated)
-                                v.animate().x(xMax.toFloat())
+                                v.animate().x(xMax)
                                     .setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
-                                    .start()
-                            else
-                                v.x = xMax.toFloat()
-                        } else {
-                            if (animated)
-                                v.animate().x(0F).setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
-                                    .start()
-                            else
-                                v.x = 0F
-                        }
-                    }
-                    Draggable.STICKY.AXIS_Y -> {
-                        if (event.rawY >= yMiddle) {
-                            if (animated)
-                                v.animate().y(yMax.toFloat())
-                                    .setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
-                                    .start()
-                            else
-                                v.y = yMax.toFloat()
-                        } else {
-                            if (animated)
-                                v.animate().y(0F)
-                                    .setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+//                                        minimizeBtnListener.onPositionChanged(v, StickyRestSide.RIGHT)
+                                    }
+                                    .setListener(object : AnimatorListenerAdapter() {
+                                        override fun onAnimationEnd(animation: Animator?) {
+                                            super.onAnimationEnd(animation)
+                                            Log.d("drg", "Animate END Sticky X RIGHT")
+                                        }
+                                    })
                                     .start()
                             else {
-                                if (animated)
-                                    v.animate().y(0F).setDuration(DURATION_MILLIS)
-                                        .setUpdateListener { draggableListener?.onPositionChanged(v) }
-                                        .start()
-                                else
-                                    v.y = 0F
+                                v.x = xMax
+//                                minimizeBtnListener.onPositionChanged(v, StickyRestSide.RIGHT)
+                            }
+                        } else {
+                            if (animated)
+                                v.animate().x(marginStart).setDuration(DURATION_MILLIS)
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+//                                        minimizeBtnListener.onPositionChanged(v, StickyRestSide.LEFT)
+                                    }
+                                    .start()
+                            else {
+                                v.x = marginStart
+//                                minimizeBtnListener.onPositionChanged(v, StickyRestSide.LEFT)
                             }
                         }
                     }
-                    Draggable.STICKY.AXIS_XY -> {
-                        if (event.rawX >= xMiddle) {
+                    DraggableView.Mode.STICKY_Y -> {
+                        if (event.rawY >= yMiddle) {
                             if (animated)
-                                v.animate().x(xMax.toFloat())
+                                v.animate().y(yMax)
                                     .setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+//                                        minimizeBtnListener.onPositionChanged(v, StickyRestSide.BOTTOM)
+                                    }
                                     .start()
                             else
-                                v.x = xMax.toFloat()
+                                v.y = yMax
                         } else {
                             if (animated)
-                                v.animate().x(0F).setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
+                                v.animate().y(marginTop)
+                                    .setDuration(DURATION_MILLIS)
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+                                    }
                                     .start()
-                            v.x = 0F
+                            else
+                                v.y = marginTop
+                        }
+                    }
+                    DraggableView.Mode.STICKY_XY -> {
+                        if (event.rawX >= xMiddle) {
+                            if (animated)
+                                v.animate().x(xMax)
+                                    .setDuration(DURATION_MILLIS)
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+                                    }
+                                    .start()
+                            else
+                                v.x = xMax
+                        } else {
+                            if (animated)
+                                v.animate().x(marginStart).setDuration(DURATION_MILLIS)
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+                                    }
+                                    .start()
+                            v.x = marginStart
                         }
 
                         if (event.rawY >= yMiddle) {
                             if (animated)
-                                v.animate().y(yMax.toFloat())
+                                v.animate().y(yMax)
                                     .setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+                                    }
                                     .start()
                             else
-                                v.y = yMax.toFloat()
+                                v.y = yMax
                         } else {
                             if (animated)
-                                v.animate().y(0F).setDuration(DURATION_MILLIS)
-                                    .setUpdateListener { draggableListener?.onPositionChanged(v) }
+                                v.animate().y(marginTop).setDuration(DURATION_MILLIS)
+                                    .setUpdateListener {
+                                        draggableListener?.onPositionChanged(v)
+                                    }
                                     .start()
                             else
-                                v.y = 0F
+                                v.y = marginTop
                         }
+                    }
+                    else -> {
                     }
                 }
 
@@ -142,6 +196,40 @@ fun View.makeDraggable(
         }
         true
     }
+}
+
+@Deprecated(
+    "Use / Setup with DraggableView class or setupDraggable() builder instead",
+    ReplaceWith(
+        "DraggableView.Builder(view).build()",
+        "io.github.hyuwah.draggableviewlib.DraggableView",
+    )
+)
+@JvmOverloads
+fun View.makeDraggable(
+    stickyAxis: Draggable.STICKY = Draggable.STICKY.NONE,
+    animated: Boolean = true,
+    draggableListener: DraggableListener? = null
+) {
+    val axisMode = when (stickyAxis) {
+        Draggable.STICKY.NONE -> DraggableView.Mode.NON_STICKY
+        Draggable.STICKY.AXIS_X -> DraggableView.Mode.STICKY_X
+        Draggable.STICKY.AXIS_Y -> DraggableView.Mode.STICKY_Y
+        Draggable.STICKY.AXIS_XY -> DraggableView.Mode.STICKY_XY
+    }
+    DraggableView.Builder(this)
+        .setStickyMode(axisMode)
+        .setAnimated(animated)
+        .setListener(draggableListener)
+        .build()
+}
+
+/**
+ * Shortcut / extension function way for DraggableView.Builder
+ * essentially the same
+ */
+fun <T : View> T.setupDraggable(): DraggableView.Builder<T> {
+    return DraggableView.Builder(this)
 }
 
 /**
